# PERF-#6478: Do not propagate new columns if they're identical to the previous ones