--- a/modin/core/dataframe/pandas/dataframe/dataframe.py
+++ b/modin/core/dataframe/pandas/dataframe/dataframe.py
@@ -581,18 +581,8 @@ class PandasDataframe(ClassLogger):
         if self.has_materialized_columns:
             # do not set new columns if they're identical to the previous ones
             if (
-                # `Index.equals()` doesn't compare metadata, thus we have to compare
-                # it manually. Here we process the simpliest and the most common case only
-                # (when index the index is a 'pandas.Index' dtype). Other cases are not that
-                # common and we can omit them
-                type(new_columns) in (pandas.Index, pandas.MultiIndex)
-                and (
-                    type(new_columns)
-                    is type(self.columns)  # noqa; here we need exact types comparison
-                )
-                and new_columns.name == self.columns.name
-                and new_columns.names == self.columns.names
-                and new_columns.equals(self.columns)
+                isinstance(new_columns, pandas.Index)
+                and self.columns.identical(new_columns)
             ) or (
                 not isinstance(new_columns, pandas.Index)
                 and np.array_equal(self.columns.values, new_columns)