<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/development/contributing.html
if you have questions about contributing.
-->

## What do these changes do?

The PR adds this new behavior for `._set_columns()` only, this is for the reason that the row indices are usually much longer and more complex than columns, thus their equality comparison may be much more expensive and would require additional heuristics valuing whether it's worth performing expensive comparison or not.

**Motivation for this PR:**

There are quite a few places where modin's implementations unconditionally overwrite the frame's columns with some specified key even if the original columns are identical to this key. For example, when setting new values for a column, the implementation sets the key as a new column value:
https://github.com/modin-project/modin/blob/3b3bba9a83b6bec35e86826800004987afe24d47/modin/core/storage_formats/pandas/query_compiler.py#L2817-L2818

New columns assignment happens despite the fact, that the old column value could be the same as the new one, for example as here:
`df["col_key"] = df["col_key"] * 2`

The same also applies to the `.insert()` operation: 
`df.insert(df2["value"], col="value")`
https://github.com/modin-project/modin/blob/3b3bba9a83b6bec35e86826800004987afe24d47/modin/core/storage_formats/pandas/query_compiler.py#L2925-L2926

There are probably more cases where this happens, those are the ones that I've seen in our workloads.

**Is skipping `._set_columns()` helps a lot?**

The `._propagate_index_objs()` call triggers empty partitions filtering that requires knowledge about row AND column partitions shapes (if they're not available, the computation will be triggered):
https://github.com/modin-project/modin/blob/3b3bba9a83b6bec35e86826800004987afe24d47/modin/core/dataframe/pandas/dataframe/dataframe.py#L709

Thus, here are two cases measured of setting new columns: if the index cache is available and when it's not available (its computation will be triggered during `._propagate_index_objs()`): 

| |master branch|this PR branch|
|-|-|-|
|with index cache (max; mean)| 0.069; 0.05 | 0.001; 0.001|
|without index cache (max; mean)| 0.12; 0.1 | 0.001; 0.001|

<details><summary>script to measure</summary>

```python
import modin.pandas as pd
import modin.config as cfg
import numpy as np
from timeit import default_timer as timer
import pandas

from asv_bench.benchmarks.utils.common import execute

# start all the workers
pd.DataFrame([np.arange(cfg.MinPartitionSize.get()) for _ in range(cfg.NPartitions.get() ** 2)]).to_numpy()
NROWS = 1_000_000
NCOLS = 1
NRUNS = 10

df = pd.DataFrame(pandas.DataFrame({f"col{i}": np.arange(NROWS) for i in range(NCOLS)}))

reses = []

for _ in range(NRUNS):
    t1 = timer()
    df.columns = df.columns.copy()
    if df._query_compiler._modin_frame._deferred_column:
        df._query_compiler._modin_frame._propagate_index_objs(axis=1)
    execute(df)
    reses.append(timer() - t1)

print(f"Results WITH index cache available: max={np.max(reses)}; mean={np.mean(reses)}")

reses = []

def reset_cache(df):
    mf = df._query_compiler._modin_frame
    mf.set_index_cache(None)
    mf._row_lengths_cache = None
    for part in mf._partitions.flatten():
        part._lengths_cache = None

for _ in range(NRUNS):
    reset_cache(df)
    t1 = timer()
    df.columns = df.columns.copy()
    if df._query_compiler._modin_frame._deferred_column:
        df._query_compiler._modin_frame._propagate_index_objs(axis=1)
    execute(df)
    reses.append(timer() - t1)

print(f"Results WITHOUT index cache available: max={np.max(reses)}; mean={np.mean(reses)}")
```

</details>

In this case, there were not delayed computations pending for the frame, in case there are delayed computations the performance improvement is expected to be even higher.

- [x] first commit message and PR title follow format outlined [here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
  > **_NOTE:_**  If you edit the PR title to match this format, you need to add another commit (even if it's empty) or amend your last commit for the CI job that checks the PR title to pick up the new PR title.
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #6478 <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->